movies_api/requirements.txt
python-dotenv
mysql-connector-python
pymysql
requests
flask-lambda-python36
flask-marshmallow
flask-sqlalchemy
marshmallow-sqlalchemy
